Complete overhaul of the software structure
The new structure is more complex, however it
allows using features, which are possibly not
available on all headsets.

For doing so, a device registry was
created, it provides (via device.h) a generic
structure for calling features and specifying
the capability of the current implementation
(or the headset).

The simple makefile was discarded in favor of
cmake.